backend/requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
blinker==1.4
CacheControl==0.12.4
cachetools==2.1.0
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
cssselect2==0.2.1
defusedxml==0.5.0
envparse==0.2.0
firebase-admin==2.10.0
Flask==1.0.2
flask-cors==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
google-api-core==1.1.2
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.9.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.7.1
honcho==1.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.0
msgpack-python==0.5.6
Pillow==5.1.0
protobuf==3.5.2.post1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.6.3
PyPDF2==1.26.0
python-dateutil==2.7.3
python-decouple==3.1
python-editor==1.0.3
pytz==2018.4
requests>=2.20.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.7
tinycss2==0.6.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
marshmallow-jsonapi
pycrypto==2.6.1
stripe==1.84.0
